[fire-hudi][feature] 调整hudi-connector隐式转换，支持从配置中获取hudi参数
