+++ fire-connectors/spark-connectors/spark-hudi/pom.xml
@@ -55,18 +55,6 @@
             <scope>provided</scope>
         </dependency>
         <dependency>
-            <groupId>com.zto.fire</groupId>
-            <artifactId>fire-spark_${spark.reference}</artifactId>
-            <version>${fire.version}</version>
-            <scope>provided</scope>
-        </dependency>
-        <dependency>
-            <groupId>com.zto.fire</groupId>
-            <artifactId>fire-enhance-spark_${spark.reference}</artifactId>
-            <version>${fire.version}</version>
-            <scope>provided</scope>
-        </dependency>
-        <dependency>
             <groupId>org.apache.hudi</groupId>
             <artifactId>hudi-common</artifactId>
             <version>${hudi.version}</version>
