+++ fire-engines/fire-spark/pom.xml
@@ -49,6 +49,12 @@
             <version>${fire.version}</version>
             <scope>${maven.scope}</scope>
         </dependency>
+        <dependency>
+            <groupId>com.zto.fire</groupId>
+            <artifactId>fire-connector-spark-hudi_${spark.reference}</artifactId>
+            <version>${fire.version}</version>
+            <scope>${maven.scope}</scope>
+        </dependency>
 
         <!-- spark相关依赖 -->
         <dependency>
